refactor signal code into its own file
